# [Relay][SimplifyInference] Express Softmax as sequence of Relay ops