@masahi 2020-02-14
please make sure you benchmark on 4d spatial inputs. In cuda softmax schedule, there is a special case handling for these inputs.