@anijain2305 2020-02-14
> Great job!
> I'm not sure whether we want to keep softmax compute & schedule though. If someone build training on top of relay ir, they might not enable simplifyinference thus need softmax as a single op.

I see. Ok, so I will only change the SimplifyInference pass but keep all the compute/schedule etc. Basically, if one disables SimplifyInference, we should still be able to run