@anijain2305 2020-02-14
> please make sure you benchmark on 4d spatial inputs. In cuda softmax schedule, there is a special case handling for these inputs.

I think you meant 2D?
https://github.com/apache/incubator-tvm/blob/master/topi/python/topi/cuda/softmax.py#L55

I see specialization for 2D softmax