@zhiics 2020-02-14
I think we may need to keep Softmax schedule as well. We can remove only if we treat them as batchnorm.

But if ones disables simplifyinference, we may still not be able to run if it has batchnorm