@masahi 2020-02-14
Correct, but look closer, it is for input dim more than 2d.

See the PR below for background. 
https://github.com/apache/incubator-tvm/pull/2338