@anijain2305 2020-02-14
I updated the original post to share more numbers. Please have a look.

@soiferj @masahi