@icemelon9 2020-02-15
Posted a comment on the discussion board. I think we should discuss more on whether we breakdown softmax to a seq of ops.